Se agrego la parte de consumos con el fitro de busqueda y el boton deshacer

+++ src/app/pages/admin/employees/viewEmploy/page.jsx
@@ -26,6 +26,12 @@ export default function EmployeeList() {
   }, []);
 
   const handleDelete = async (id) => {
+    const isConfirmed = confirm('¿Estás seguro de que deseas eliminar este empleado?');
+    
+    if (!isConfirmed) {
+      return;
+    }
+
     try {
       const res = await fetch(`/api/employs/${id}`, {
         method: 'DELETE',
